Remove GitHub Actions workflow for Terraform
This commit removes the `terraform.yml` file from the `.github/workflows` directory, which was used to configure a GitHub Actions workflow for deploying Terraform code changes to AWS. The workflow was triggered on push and pull request events to the `main` and `stage` branches, and used environment variables and secrets to authenticate with AWS and store the Terraform state in an S3 bucket. The workflow consisted of several steps, including checking out the source code, setting up Terraform, initializing the backend, formatting and validating the code, and creating a plan for applying the changes. However, this file has been deleted in this commit.